Move Python wrappers to separate module
